chore: restructure integration tests
