+++ compiler/integration-tests/circuits/main/Nargo.toml
