+++ compiler/integration-tests/circuits/recursion/Nargo.toml
