+++ compiler/integration-tests/test/browser/compile_prove_verify.test.ts
@@ -1,5 +1,5 @@
 import { expect } from '@esm-bundle/chai';
-import { TEST_LOG_LEVEL } from '../../environment.js';
+import { TEST_LOG_LEVEL } from '../environment.js';
 import { Logger } from 'tslog';
 import { initializeResolver } from '@noir-lang/source-resolver';
 import newCompiler, { compile, init_log_level as compilerLogLevel } from '@noir-lang/noir_wasm';
@@ -7,6 +7,8 @@ import { acvm, abi, Noir } from '@noir-lang/noir_js';
 import { BarretenbergBackend } from '@noir-lang/backend_barretenberg';
 import { ethers } from 'ethers';
 import * as TOML from 'smol-toml';
+import { getFile } from './utils.js';
+import { separatePublicInputsFromProof } from '../shared/proof.js';
 
 const provider = new ethers.JsonRpcProvider('http://localhost:8545');
 const logger = new Logger({ name: 'test', minLevel: TEST_LOG_LEVEL });
@@ -19,17 +21,6 @@ await newABICoder();
 await initACVM();
 
 compilerLogLevel('INFO');
-
-async function getFile(file_path: string): Promise<string> {
-  const file_url = new URL(file_path, import.meta.url);
-  const response = await fetch(file_url);
-
-  if (!response.ok) throw new Error('Network response was not OK');
-
-  return await response.text();
-}
-
-const FIELD_ELEMENT_BYTES = 32;
 
 const test_cases = [
   {
@@ -58,22 +49,6 @@ async function getCircuit(noirSource: string) {
   });
 
   return compile({});
-}
-
-function separatePublicInputsFromProof(
-  proof: Uint8Array,
-  numPublicInputs: number,
-): { proof: Uint8Array; publicInputs: Uint8Array[] } {
-  const publicInputs = Array.from({ length: numPublicInputs }, (_, i) => {
-    const offset = i * FIELD_ELEMENT_BYTES;
-    return proof.slice(offset, offset + FIELD_ELEMENT_BYTES);
-  });
-  const slicedProof = proof.slice(numPublicInputs * FIELD_ELEMENT_BYTES);
-
-  return {
-    proof: slicedProof,
-    publicInputs,
-  };
 }
 
 test_cases.forEach((testInfo) => {
