+++ compiler/integration-tests/test/browser/utils.ts
@@ -1,0 +1,8 @@
+export async function getFile(file_path: string): Promise<string> {
+  const file_url = new URL(file_path, import.meta.url);
+  const response = await fetch(file_url);
+
+  if (!response.ok) throw new Error('Network response was not OK');
+
+  return await response.text();
+}
