--- compiler/integration-tests/test/index.d.ts
@@ -1,1 +1,0 @@
-declare module '@aztec/bb.js';
