+++ compiler/integration-tests/web-test-runner.config.mjs
@@ -1,19 +1,19 @@
-import { defaultReporter } from "@web/test-runner";
-import { summaryReporter } from "@web/test-runner";
-import { fileURLToPath } from "url";
-import { esbuildPlugin } from "@web/dev-server-esbuild";
-import { webdriverLauncher } from "@web/test-runner-webdriver";
+import { defaultReporter } from '@web/test-runner';
+import { summaryReporter } from '@web/test-runner';
+import { fileURLToPath } from 'url';
+import { esbuildPlugin } from '@web/dev-server-esbuild';
+import { webdriverLauncher } from '@web/test-runner-webdriver';
 
 let reporter = summaryReporter();
 const debugPlugins = [];
 // eslint-disable-next-line no-undef
-if (process.env.CI !== "true" || process.env.RUNNER_DEBUG === "1") {
+if (process.env.CI !== 'true' || process.env.RUNNER_DEBUG === '1') {
   reporter = defaultReporter();
   debugPlugins.push({
-    name: "environment",
+    name: 'environment',
     serve(context) {
-      if (context.path === "/compiler/integration-tests/test/environment.js") {
-        return "export const TEST_LOG_LEVEL = 2;";
+      if (context.path === '/compiler/integration-tests/test/environment.js') {
+        return 'export const TEST_LOG_LEVEL = 2;';
       }
     },
   });
@@ -22,11 +22,11 @@ if (process.env.CI !== "true" || process.env.RUNNER_DEBUG === "1") {
 export default {
   browsers: [
     webdriverLauncher({
-      automationProtocol: "webdriver",
+      automationProtocol: 'webdriver',
       capabilities: {
-        browserName: "firefox",
-        "moz:firefoxOptions": {
-          args: ["-headless"],
+        browserName: 'firefox',
+        'moz:firefoxOptions': {
+          args: ['-headless'],
         },
       },
     }),
@@ -37,15 +37,15 @@ export default {
     }),
     ...debugPlugins,
   ],
-  files: ["test/integration/browser/**/*.test.ts"],
+  files: ['test/browser/**/*.test.ts'],
   nodeResolve: { browser: true },
   testFramework: {
     config: {
-      ui: "bdd",
+      ui: 'bdd',
     },
   },
   // eslint-disable-next-line no-undef
-  rootDir: fileURLToPath(new URL("./../..", import.meta.url)),
+  rootDir: fileURLToPath(new URL('./../..', import.meta.url)),
   testsFinishTimeout: 60 * 20e3, // 20 minutes
   reporters: [reporter],
 };
